[REFACTOR] Moved the filter UI to separate screen
